pyproject.toml
[tool.poetry]
name = "robusta_actions"
version = "0.0.1"
description = ""
authors = ["Arik Alon <arikalon1@users.noreply.github.com>"]

[tool.poetry.dependencies]
kubernetes = "^12.0.1"

[tool.poetry.dev-dependencies]
robusta-cli = "^0.9.17"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
